Fix broken frontend structure
