+++ frontend/src/components/auth/AuthShared.jsx
