+++ frontend/src/index.css
@@ -1,68 +1,28 @@
+@import url('https://fonts.googleapis.com/css2?family=JetBrains+Mono:wght@400;600;700&family=Sora:wght@300;400;500;600;700&display=swap');
+@import 'tailwindcss';
+
+/* ── Design tokens ── */
 :root {
-  font-family: system-ui, Avenir, Helvetica, Arial, sans-serif;
-  line-height: 1.5;
-  font-weight: 400;
-
-  color-scheme: light dark;
-  color: rgba(255, 255, 255, 0.87);
-  background-color: #242424;
-
-  font-synthesis: none;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-a {
-  font-weight: 500;
-  color: #646cff;
-  text-decoration: inherit;
-}
-a:hover {
-  color: #535bf2;
+  --rt-bg:            #0d1117;
+  --rt-surface:       #161b22;
+  --rt-surface2:      #1c2333;
+  --rt-border:        #30363d;
+  --rt-border-bright: #444c56;
+  --rt-text:          #e6edf3;
+  --rt-muted:         #7d8590;
+  --rt-dim:           #484f58;
+  --rt-flame:         #f78166;
+  --rt-sky:           #79c0ff;
+  --rt-leaf:          #3fb950;
+  --rt-amber:         #d29922;
+  --rt-ember:         #f85149;
+  --rt-orchid:        #bc8cff;
 }
 
 body {
   margin: 0;
-  display: flex;
-  place-items: center;
-  min-width: 320px;
   min-height: 100vh;
-}
-
-h1 {
-  font-size: 3.2em;
-  line-height: 1.1;
-}
-
-button {
-  border-radius: 8px;
-  border: 1px solid transparent;
-  padding: 0.6em 1.2em;
-  font-size: 1em;
-  font-weight: 500;
-  font-family: inherit;
-  background-color: #1a1a1a;
-  cursor: pointer;
-  transition: border-color 0.25s;
-}
-button:hover {
-  border-color: #646cff;
-}
-button:focus,
-button:focus-visible {
-  outline: 4px auto -webkit-focus-ring-color;
-}
-
-@media (prefers-color-scheme: light) {
-  :root {
-    color: #213547;
-    background-color: #ffffff;
-  }
-  a:hover {
-    color: #747bff;
-  }
-  button {
-    background-color: #f9f9f9;
-  }
+  background-color: var(--rt-bg);
+  color: var(--rt-text);
+  font-family: 'Sora', ui-sans-serif, system-ui, sans-serif;
 }
