+++ frontend/src/pages/HowItWorks.jsx
