+++ frontend/src/pages/ProjectScopeManager.jsx
