+++ frontend/src/pages/ProjectWorkspace.jsx
