+++ frontend/src/pages/ScopeConfig.jsx
