+++ frontend/src/pages/TermsModal.jsx
