--- frontend/vitest.config.js
@@ -1,9 +1,0 @@
-import { defineConfig } from 'vitest/config'
- 
-export default defineConfig({
-  test: {
-    include: ['web/src/**/*.test.js'],
-    environment: 'node',
-  },
-})
-
