--- frontend/web/src/index.css
@@ -1,28 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=JetBrains+Mono:wght@400;600;700&family=Sora:wght@300;400;500;600;700&display=swap');
-@import 'tailwindcss';
-
-/* ── Design tokens ── */
-:root {
-  --rt-bg:            #0d1117;
-  --rt-surface:       #161b22;
-  --rt-surface2:      #1c2333;
-  --rt-border:        #30363d;
-  --rt-border-bright: #444c56;
-  --rt-text:          #e6edf3;
-  --rt-muted:         #7d8590;
-  --rt-dim:           #484f58;
-  --rt-flame:         #f78166;
-  --rt-sky:           #79c0ff;
-  --rt-leaf:          #3fb950;
-  --rt-amber:         #d29922;
-  --rt-ember:         #f85149;
-  --rt-orchid:        #bc8cff;
-}
-
-body {
-  margin: 0;
-  min-height: 100vh;
-  background-color: var(--rt-bg);
-  color: var(--rt-text);
-  font-family: 'Sora', ui-sans-serif, system-ui, sans-serif;
-}
